+++ app/src/main/java/com/andyisdope/cryptowatcher/model/Transaction.kt
@@ -9,29 +9,29 @@ import android.os.Parcelable
 @Entity(tableName = "Transactions")
 data class Transaction(@PrimaryKey() val Date: Long,
                   @ColumnInfo(name = "Coin") val Coin: String,
-                  @ColumnInfo(name = "Amount") val Amount: Float,
+                  @ColumnInfo(name = "Amount") val Amount: Double,
                   @ColumnInfo(name = "Buy") val Buy: Boolean,
                   @ColumnInfo(name = "Sell") val Sell: Boolean,
-                  @ColumnInfo(name = "Price") val Price: Float,
-                  @ColumnInfo(name = "Net") val Net: Float) : Parcelable {
+                  @ColumnInfo(name = "Price") val Price: Double,
+                  @ColumnInfo(name = "Net") val Net: Double) : Parcelable {
     constructor(parcel: Parcel) : this(
             parcel.readLong(),
             parcel.readString(),
-            parcel.readFloat(),
+            parcel.readDouble(),
             parcel.readByte() != 0.toByte(),
             parcel.readByte() != 0.toByte(),
-            parcel.readFloat(),
-            parcel.readFloat()) {
+            parcel.readDouble(),
+            parcel.readDouble()) {
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeLong(Date)
         parcel.writeString(Coin)
-        parcel.writeFloat(Amount)
+        parcel.writeDouble(Amount)
         parcel.writeByte(if (Buy) 1 else 0)
         parcel.writeByte(if (Sell) 1 else 0)
-        parcel.writeFloat(Price)
-        parcel.writeFloat(Net)
+        parcel.writeDouble(Price)
+        parcel.writeDouble(Net)
     }
 
     override fun describeContents(): Int {
